Move eclipse modules to editor
- So we can reuse them in other integrations as vscode
