+++ packages/editor/src/open-dataclass/di.config.ts
@@ -1,6 +1,6 @@
 import { ContainerModule } from 'inversify';
 import { TYPES } from '@eclipse-glsp/client';
-import { OpenQuickOutlineKeyListener } from './open-data-class';
+import { OpenQuickOutlineKeyListener } from './open-dataclass';
 
 const ivyOpenDataClassModule = new ContainerModule(bind => {
   bind(TYPES.KeyListener).to(OpenQuickOutlineKeyListener);
