+++ packages/editor/src/open-inscription/di.config.ts
@@ -1,4 +1,3 @@
-import { IVY_TYPES } from '@axonivy/process-editor';
 import { ContainerModule } from 'inversify';
 import { configureActionHandler, TYPES } from '@eclipse-glsp/client';
 import { OpenAction } from 'sprotty-protocol';
@@ -6,6 +5,7 @@ import { OpenAction } from 'sprotty-protocol';
 import { OpenInscriptionMouseListener } from './mouse-listener';
 import { OpenInscriptionActionHandler, OpenInscriptionKeyListener } from './open-inscription-handler';
 import { InscribeQuickActionProvider } from './quick-action';
+import { IVY_TYPES } from '../types';
 
 const ivyOpenInscriptionModule = new ContainerModule((bind, _unbind, isBound) => {
   bind(TYPES.KeyListener).to(OpenInscriptionKeyListener);
