+++ packages/editor/src/open-inscription/mouse-listener.ts
