+++ packages/editor/src/open-inscription/quick-action.ts
@@ -1,8 +1,9 @@
 import { isOpenable, SModelElement } from '@eclipse-glsp/client';
-import { QuickAction, QuickActionLocation, SingleQuickActionProvider, StreamlineIcons } from '@axonivy/process-editor';
 import { injectable } from 'inversify';
 import { OpenAction } from 'sprotty-protocol';
 import { KeyCode } from 'sprotty/lib/utils/keyboard';
+import { StreamlineIcons } from '../StreamlineIcons';
+import { QuickAction, QuickActionLocation, SingleQuickActionProvider } from '../ui-tools/quick-action/quick-action';
 
 @injectable()
 export class InscribeQuickActionProvider extends SingleQuickActionProvider {
