+++ packages/editor/src/open-source/di.config.ts
@@ -1,7 +1,7 @@
-import { IVY_TYPES } from '@axonivy/process-editor';
 import { ContainerModule } from 'inversify';
 
 import { GoToSourceQuickActionProvider } from './quick-action';
+import { IVY_TYPES } from '../types';
 
 const ivyGoToSourceModule = new ContainerModule(bind => {
   bind(IVY_TYPES.QuickActionProvider).to(GoToSourceQuickActionProvider);
