+++ packages/editor/src/start-action/di.config.ts
@@ -1,10 +1,10 @@
-import { IVY_TYPES } from '@axonivy/process-editor';
 import { ContainerModule } from 'inversify';
 import { SearchProcessCallersActionProvider, StarProcessQuickActionProvider } from './actions';
+import { IVY_TYPES } from '../types';
 
-const ivyEditorActionModule = new ContainerModule(bind => {
+const ivyStartActionModule = new ContainerModule(bind => {
   bind(IVY_TYPES.QuickActionProvider).to(StarProcessQuickActionProvider);
   bind(IVY_TYPES.QuickActionProvider).to(SearchProcessCallersActionProvider);
 });
 
-export default ivyEditorActionModule;
+export default ivyStartActionModule;
